perf(ds): 去掉无用的强制转换

+++ src/com/cphayim/test/queue/QueuePerfTest.java
@@ -36,7 +36,7 @@ public class QueuePerfTest {
 
         long endTime = System.nanoTime();
 
-        return (endTime - startTime) / (double) 1e9;
+        return (endTime - startTime) / 1e9;
     }
 
     public static void main(String[] args) {
